requirements.txt
streamlit==0.60.0
opencv_python==4.2.0.34
numpy==1.18.5
Pillow==7.1.2
